requirements.txt
streamlit
pandas
openpyxl
scikit-learn
google-generativeai
python-dotenv